+++ tests/assertions/assert_cookies.test.js
@@ -1,0 +1,71 @@
+"use strict";
+
+const Wendigo = require('../../lib/wendigo');
+const configUrls = require('../config.json').urls;
+const utils = require('../utils');
+
+describe("Assert Cookie", function() {
+    this.timeout(5000);
+    let browser;
+
+    before(async () => {
+        browser = await Wendigo.createBrowser();
+    });
+
+    beforeEach(async () => {
+        await browser.open(configUrls.storage);
+    });
+
+    afterEach(async () => {
+        await browser.localStorage.clear();
+        await browser.cookies.clear();
+    });
+
+    it("Assert Cookie", async () => {
+        await browser.assert.cookie("username");
+        await browser.assert.cookie("username", "arthur_dent");
+    });
+
+    it("Assert Cookie Throws", async () => {
+        await utils.assertThrowsAssertionAsync(async () => {
+            await browser.assert.cookie("not-exists");
+        }, `Expected cookie "not-exists" to exist.`);
+    });
+
+    it("Assert Cookie Throws Invalid Value", async () => {
+        await utils.assertThrowsAssertionAsync(async () => {
+            await browser.assert.cookie("username", "marvin");
+        }, `Expected cookie "username" to have value "marvin", "arthur_dent" found.`);
+    });
+
+    it("Assert Cookie Throws Custom Message", async () => {
+        await utils.assertThrowsAssertionAsync(async () => {
+            await browser.assert.cookie("username", "marvin", "cookie fails");
+        }, `cookie fails`);
+    });
+
+    it("Assert Not Cookie", async () => {
+        await browser.assert.not.cookie("not-exists");
+        await browser.assert.not.cookie("username", "marvin");
+    });
+
+    it("Assert Not Cookie Throws", async () => {
+        await utils.assertThrowsAssertionAsync(async () => {
+            await browser.assert.not.cookie("username");
+        }, `Expected cookie "username" to not exist.`);
+    });
+
+    it("Assert Not Cookie Throws Invalid Value", async () => {
+        await utils.assertThrowsAssertionAsync(async () => {
+            await browser.assert.not.cookie("username", "arthur_dent");
+        }, `Expected cookie "username" to not have value "arthur_dent".`);
+    });
+
+    it("Assert Not Cookie Throws Custom Message", async () => {
+        await utils.assertThrowsAssertionAsync(async () => {
+            await browser.assert.not.cookie("username", "arthur_dent", "not cookie fails");
+        }, `not cookie fails`);
+    });
+
+
+});